% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl}
\alias{crawl}
\title{Crawl}
\usage{
crawl()
}
\arguments{
\item{data}{Reactive function brings back all the Comments with selected rating}
}
\value{
Brings back all the comments with selected rating
}
\description{
Crawl
}
